+++ packages/components/src/Form/Inputs/Combobox/ComboboxList.tsx
@@ -51,7 +51,7 @@ import once from 'lodash/once'
 import throttle from 'lodash/throttle'
 import { usePopover } from '../../../Popover'
 import { useResize } from '../../../utils'
-import { ComboboxOptionIndicatorProps } from './ComboboxOptionIndicator'
+import { ComboboxOptionIndicatorProps } from './types'
 import { ComboboxContext, ComboboxMultiContext } from './ComboboxContext'
 import { useBlur } from './utils/useBlur'
 import { useKeyDown } from './utils/useKeyDown'
