+++ packages/components/src/Form/Inputs/Combobox/index.ts
@@ -33,4 +33,5 @@ export * from './ComboboxContext'
 export * from './ComboboxInput'
 export * from './ComboboxList'
 export * from './ComboboxOption'
+export * from './types'
 export * from './utils'
