+++ packages/components/src/Form/Inputs/Select/index.ts
@@ -28,3 +28,4 @@ export * from './Select'
 export * from './SelectInputIcon'
 export * from './SelectMulti'
 export * from './SelectOptions'
+export * from './types'
